Split components into separate files.
